requirements.txt
flask
python-telegram-bot==20.3
python-dotenv
requests
telethon
